Use a dictionary for the duplicate files list.
It will run a select query to find rows with equivalent hashes for the data and the key is the name of the file that was flagged as a duplicate.

At this point, it will print out the raw dictionary, but maybe someday I will write code to make it a bit prettier.