Add a few features/fixes.
Create a unique index for the hash and filename.

Check if values is a list or tuple before trying to execute the query in execquerycommit.

Add code to lowercase the inferred table name and an argument to toggle it back to the old behavior.

Annotate the "base" variable in infertableadd() to pathlib.Path