Some modifications to cleantablename()
Also, run cleantablename() on table name specified on the commandline.

Added another parameter to cleantablename() so that whether it lowercases or not is determined by this new parameter and not args.lower directly.
Reasoning being that if a table name is specified on the command line, we don't want a sanitized version of what we put in, without forcing a lower-cased version if you don't put in a specific commandline parameter.