Add a (possibly unnecessary, but better safe than sorry) check to make sure that args.autovacuum is set before trying to assess it's state.
Make some formatting changes.

Split the choices list for --autovacuum into 3 dict keys and combine them back into a single dict key with the same name. This is a readability change.

Add some checks to make sure that the autovacuum pragma was actually changed if a different autovacuum mode is requested.

If it is the case that the autovacuum mode is not changed, it will return false so that a vacuum is not run unnecessarily.