Code cleanup, more below.
Move changing the text factory function to the extract function since that's the only place it's used and it was already hidden behind an if statement for extract mode anyway.

Calculate self.files differently for extract mode since the list comprehension that was in place only makes sense for add mode. Said list comprehension is still used for add mode.

Add an execquerymanynocommit function. Like its commit counterpart, it's not used at this point. It was going to be used for extracting files faster, but apparently either executemany doesn't work with select statements or I'm not crafting the query correctly, so I'm keeping the code around.

Switch some uses of args.files with self.files in calcextractquery. This was how I was able to determine that the self.files list comprehension was inappropriate for extract mode. Also, only calculate the length of self.files once in calcextractquery() and use that variable to calculate which query should be used.

Absolutely make sure that the execquerynocommit statement returns data and doesn't somehow switch to the "action query" mode.