Clean up the infertable extract mode.
The if statements were wrong. the cleantablename() function should be run on out no matter what if out is specified. The files[0] if statement already doesn't run if out is specified.

This might also fix a bug where infertable was failing in extract mode, not tested at time of commit.