Added some hacks to FINALLY get file exclusion to work.
It only works on file names at this point because i'm using pathlib to remove all directory components from the list entries.

I hope to figure out how to be able to exclude a file in a particular directory at some point, but that is a project for the future.