Add globlist2() and remove a corner case in globlist()
Adding globlist2() as an experiment in cleaning up and removing complexity from globlist(). It will not behave any different if the main script is in add or extract mode. Extract mode doesn't even use the glob translation features of globlist anyway.

Removed a corner case where no code would be run in globlist(). Theoretically, it shouldn't happen, but since this function could be used elsewhere, it makes no sense to have code not run if mode is not one of add or extract. It defaults to add mode.